added components for answer types
